Use compiler-libs, handle multiple OCaml version, source info.
1. Decide to use compiler-libs that comes with the compiler
rather than ocaml-compiler-libs (from Janestreet) as it
is always available when the compiler is.
2. Use virtual library interface to printtyp, to be able to
handle multiple OCaml versions.
3. Add start of support for showing source when dumping
code. Still need to setup loadpath, but it works if you
call the dumper from the workspace root. To provide this
support I copied (and sometimes edited) the modules under
the debugger directory, that were taken from ocaml/debugger.